+++ examples/json_consumer_encryption_example/json_consumer_encryption_example.go
@@ -31,6 +31,7 @@ import (
 	"github.com/confluentinc/confluent-kafka-go/v2/schemaregistry/rules/encryption/azurekms"
 	"github.com/confluentinc/confluent-kafka-go/v2/schemaregistry/rules/encryption/gcpkms"
 	"github.com/confluentinc/confluent-kafka-go/v2/schemaregistry/rules/encryption/hcvault"
+	"github.com/confluentinc/confluent-kafka-go/v2/schemaregistry/rules/encryption/localkms"
 
 	"github.com/confluentinc/confluent-kafka-go/v2/kafka"
 	"github.com/confluentinc/confluent-kafka-go/v2/schemaregistry"
@@ -51,6 +52,7 @@ func main() {
 	azurekms.Register()
 	gcpkms.Register()
 	hcvault.Register()
+	localkms.Register()
 	encryption.Register()
 
 	bootstrapServers := os.Args[1]