+++ schemaregistry/schemaregistry_client.go
@@ -31,9 +31,15 @@ import (
 
 /* Schema Registry API endpoints
 *
+* ====Contexts====
+* Fetch JSON array str:context of all contexts
+* -GET /contexts returns: JSON array string: contexts; raises: 500[01]
+*
 * ====Schemas====
 * Fetch string: schema(escaped) identified by the input id.
 * -GET /schemas/ids/{int: id} returns: JSON blob: schema; raises: 404[03], 500[01]
+* Fetch string: JSON array (subject, version) of schemas identified by ID.
+* -GET /schemas/ids/{int: id}/versions returns: JSON array; raises: 404[03], 500[01]
 *
 * ====Subjects====
 * Fetch JSON array str:subject of all registered subjects
@@ -188,6 +194,12 @@ type SchemaInfo struct {
 	References []Reference `json:"references,omitempty"`
 	Metadata   *Metadata   `json:"metadata,omitempty"`
 	RuleSet    *RuleSet    `json:"ruleSet,omitempty"`
+}
+
+// SubjectAndVersion represents a pair of subject and version
+type SubjectAndVersion struct {
+	Subject string `json:"subject,omitempty"`
+	Version int    `json:"version,omitempty"`
 }
 
 // MarshalJSON implements the json.Marshaler interface
@@ -350,10 +362,12 @@ var _ Client = new(client)
 // The Schema Registry's REST interface is further explained in Confluent's Schema Registry API documentation
 // https://github.com/confluentinc/schema-registry/blob/master/client/src/main/java/io/confluent/kafka/schemaregistry/client/SchemaRegistryClient.java
 type Client interface {
+	GetAllContexts() ([]string, error)
 	Config() *Config
 	Register(subject string, schema SchemaInfo, normalize bool) (id int, err error)
 	RegisterFullResponse(subject string, schema SchemaInfo, normalize bool) (result SchemaMetadata, err error)
 	GetBySubjectAndID(subject string, id int) (schema SchemaInfo, err error)
+	GetSubjectsAndVersionsByID(id int) (subjectAndVersion []SubjectAndVersion, err error)
 	GetID(subject string, schema SchemaInfo, normalize bool) (id int, err error)
 	GetLatestSchemaMetadata(subject string) (SchemaMetadata, error)
 	GetSchemaMetadata(subject string, version int) (SchemaMetadata, error)
@@ -459,6 +473,14 @@ func NewClient(conf *Config) (Client, error) {
 	return handle, nil
 }
 
+// Returns a string slice containing all available contexts
+func (c *client) GetAllContexts() ([]string, error) {
+	var result []string
+	err := c.restService.HandleRequest(internal.NewRequest("GET", context, nil), &result)
+
+	return result, err
+}
+
 // Config returns the client config
 func (c *client) Config() *Config {
 	return c.config
@@ -548,6 +570,14 @@ func (c *client) GetBySubjectAndID(subject string, id int) (schema SchemaInfo, e
 	}
 	c.idToSchemaInfoCacheLock.Unlock()
 	return *newInfo, err
+}
+
+// GetSubjectsAndVersionsByID returns the subject-version pairs for a given ID.
+// Returns SubjectAndVersion object on success.
+// This method cannot not use caching to increase performance.
+func (c *client) GetSubjectsAndVersionsByID(id int) (subbjectsAndVersions []SubjectAndVersion, err error) {
+	err = c.restService.HandleRequest(internal.NewRequest("GET", subjectsAndVersionsByID, nil, id), &subbjectsAndVersions)
+	return
 }
 
 // GetID checks if a schema has been registered with the subject. Returns ID if the registration can be found
@@ -973,7 +1003,7 @@ func (c *client) GetConfig(subject string, defaultToGlobal bool) (result ServerC
 	return result, err
 }
 
-// UpdateCompatibility updates subject's config
+// UpdateConfig updates subject's config
 // Returns new config string upon success
 func (c *client) UpdateConfig(subject string, update ServerConfig) (result ServerConfig, err error) {
 	err = c.restService.HandleRequest(internal.NewRequest("PUT", internal.SubjectConfig, &update, url.PathEscape(subject)), &result)