+++ schemaregistry/serde/avro/avro.go
@@ -29,11 +29,7 @@ func resolveAvroReferences(c schemaregistry.Client, schema schemaregistry.Schema
 		if err != nil {
 			return nil, err
 		}
-		info := schemaregistry.SchemaInfo{
-			Schema:     metadata.Schema,
-			SchemaType: metadata.SchemaType,
-			References: metadata.References,
-		}
+		info := metadata.SchemaInfo
 		_, err = resolveAvroReferences(c, info, ns)
 		if err != nil {
 			return nil, err